Profile loading state

+++ APIs/deleteUser.ts
@@ -1,32 +1,29 @@
 import { fetchAuthSession } from "aws-amplify/auth";
 import axios from "axios";
-
-const INVOKE_URL = process.env.INVOKE_URL;
-const url = `${INVOKE_URL}/customer/user/delete`;
-
+import Constants from "expo-constants";
+const { apiPath } = Constants.expoConfig?.extra || {};
+const url = apiPath + "/customer/user/delete";
 const deleteUser = async () => {
-    try {
-        const { tokens } = await fetchAuthSession();
-        const accessToken = tokens?.idToken;
-        if (!accessToken) {
-            throw new Error("No access token available");
-        }
-
-        const { data } = await axios.delete(url, {
-            headers: {
-                "Content-Type": "application/json",
-                "Authorization": `Bearer ${accessToken}`,
-            },
-        });
-
-        return data;
-    } catch (error: any) {
-        console.error("Error deleting user:", error.response?.data || error.message);
-        console.error("Status code:", error.response?.status);
-        console.error("Headers:", error.response?.headers);
-
-        return null;
+  try {
+    const { tokens } = await fetchAuthSession();
+    const accessToken = tokens?.idToken;
+    if (!accessToken) {
+      throw new Error("No access token available");
     }
+    if (!url) {
+      throw new Error("CUSTOMER_DELETE_ENDPOINT is not defined");
+    }
+    const { data } = await axios.delete(url, {
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${accessToken}`,
+      },
+    });
+
+    return data;
+  } catch (error: any) {
+    return null;
+  }
 };
 
 export default deleteUser;
